// src/lib/typings/ICity.ts
export interface ICity {
  city: string;
  country: string;
  countryCode: string;
  id: number;
  latitude: number;
  longitude: number;
  name: string;
  population: number;
  region: string;
  regionCode: string;
  regionWdId: string;
  type: string;
  wikiDataId: string;
}

export interface ICityResponse {
  data: ICity[];
  links: {
    href: string;
    rel: string;
  }[];
  metadata: {
    currentOffset: number;
    totalCount: number;
  };
}